projetos/data_science/bots/habitflow_telegram/requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
google-cloud-storage==3.8.0
pytz